backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["ogata katsuya"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = "^2.0.30"
aiomysql = "^0.2.0"
aiohttp-cors = "^0.7.0"
passlib = "^1.7.4"
pydantic = "^2.7.1"
python-jose = "^3.3.0"
psycopg2 = "^2.9.9"
pytest-mock = "^3.14.0"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.23.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
